fix #7942 feat(nimbus): Move timeline up
