+++ app/experimenter/nimbus-ui/src/components/PageSummary/index.tsx
@@ -4,6 +4,7 @@
 
 import { RouteComponentProps } from "@reach/router";
 import React, { useContext, useMemo, useState } from "react";
+import { Badge } from "react-bootstrap";
 import Alert from "react-bootstrap/Alert";
 import { useChangeOperationMutation, useReviewCheck } from "../../hooks";
 import { ReactComponent as InfoCircle } from "../../images/info-circle.svg";
@@ -14,6 +15,7 @@ import {
 } from "../../lib/constants";
 import { ExperimentContext } from "../../lib/contexts";
 import { getStatus, getSummaryAction } from "../../lib/experiment";
+import { getExperiment_experimentBySlug } from "../../types/getExperiment";
 import {
   NimbusExperimentPublishStatusEnum,
   NimbusExperimentStatusEnum,
@@ -22,6 +24,7 @@ import AppLayoutWithExperiment from "../AppLayoutWithExperiment";
 import ChangeApprovalOperations from "../ChangeApprovalOperations";
 import Head from "../Head";
 import Summary from "../Summary";
+import SummaryTimeline from "../Summary/SummaryTimeline";
 import FormLaunchDraftToPreview from "./FormLaunchDraftToPreview";
 import FormLaunchDraftToReview from "./FormLaunchDraftToReview";
 import FormLaunchPreviewToReview from "./FormLaunchPreviewToReview";
@@ -175,6 +178,12 @@ const PageSummary = (props: RouteComponentProps) => {
   return (
     <AppLayoutWithExperiment testId="PageSummary" setHead={false}>
       <Head title={`${experiment.name} – ${summaryTitle}`} />
+      <h5 className="mb-3">
+        Timeline
+        {status.live && <StatusPills {...{ experiment }} />}
+      </h5>
+
+      <SummaryTimeline {...{ experiment }} />
 
       {submitError && (
         <Alert data-testid="submit-error" variant="warning">
@@ -183,9 +192,9 @@ const PageSummary = (props: RouteComponentProps) => {
       )}
 
       {summaryAction && (
-        <h2 className="mt-3 mb-4 h4">
+        <h5 className="mt-3 mb-4 ml-3">
           {summaryAction} {launchDocs}
-        </h2>
+        </h5>
       )}
 
       <ChangeApprovalOperations
@@ -245,3 +254,33 @@ const PageSummary = (props: RouteComponentProps) => {
 };
 
 export default PageSummary;
+
+const StatusPills = ({
+  experiment,
+}: {
+  experiment: getExperiment_experimentBySlug;
+}) => (
+  <>
+    {experiment.isEnrollmentPaused === false && (
+      <StatusPill
+        testId="pill-enrolling-active"
+        label="Enrolling Users in Progress"
+      />
+    )}
+    {experiment.isEnrollmentPaused && experiment.enrollmentEndDate && (
+      <StatusPill
+        testId="pill-enrolling-complete"
+        label="Enrollment Complete"
+      />
+    )}
+  </>
+);
+
+const StatusPill = ({ label, testId }: { label: string; testId: string }) => (
+  <Badge
+    className="ml-2 border rounded-pill px-2 bg-white border-primary text-primary font-weight-normal"
+    data-testid={testId}
+  >
+    {label}
+  </Badge>
+);
