+++ app/experimenter/nimbus-ui/src/components/Summary/index.test.tsx
@@ -15,7 +15,7 @@ import { createMutationMock, Subject } from "./mocks";
 describe("Summary", () => {
   it("renders expected components", () => {
     render(<Subject />);
-    expect(screen.getByTestId("summary-timeline")).toBeInTheDocument();
+
     expect(screen.queryByTestId("experiment-end")).not.toBeInTheDocument();
     expect(
       screen.queryByTestId("link-monitoring-dashboard"),
@@ -62,31 +62,6 @@ describe("Summary", () => {
     expect(
       screen.queryByTestId("summary-page-signoff-not-launched"),
     ).toBeInTheDocument();
-  });
-
-  it("renders enrollment active badge if enrollment is not paused", async () => {
-    render(
-      <Subject
-        props={{
-          status: NimbusExperimentStatusEnum.LIVE,
-          isEnrollmentPaused: false,
-        }}
-      />,
-    );
-    await screen.findByTestId("pill-enrolling-active");
-  });
-
-  it("renders enrollment complete badge if enrollment is not paused", async () => {
-    render(
-      <Subject
-        props={{
-          status: NimbusExperimentStatusEnum.LIVE,
-          isEnrollmentPaused: true,
-          enrollmentEndDate: new Date().toISOString(),
-        }}
-      />,
-    );
-    await screen.findByTestId("pill-enrolling-complete");
   });
 
   it("renders the end experiment button if the experiment is live and idle", async () => {
@@ -190,36 +165,6 @@ describe("Summary", () => {
         expect(refetch).toHaveBeenCalled();
         expect(screen.queryByTestId("submit-error")).not.toBeInTheDocument();
       });
-    });
-
-    it("verify cancel review doesn't change enrollment days", async () => {
-      const refetch = jest.fn();
-      const { experiment } = mockExperimentQuery("demo-slug", {
-        status: NimbusExperimentStatusEnum.LIVE,
-        publishStatus: NimbusExperimentPublishStatusEnum.REVIEW,
-      });
-      const mutationMock = createMutationMock(
-        experiment.id!,
-        NimbusExperimentPublishStatusEnum.IDLE,
-        {
-          statusNext: NimbusExperimentStatusEnum.LIVE,
-          changelogMessage: CHANGELOG_MESSAGES.CANCEL_REVIEW,
-          isEnrollmentPaused: false,
-        },
-      );
-      render(
-        <Subject props={experiment} mocks={[mutationMock]} {...{ refetch }} />,
-      );
-
-      await screen.findByTestId("cancel-review-start");
-      fireEvent.click(screen.getByTestId("cancel-review-start"));
-      await waitFor(() => {
-        expect(refetch).toHaveBeenCalled();
-        expect(screen.queryByTestId("submit-error")).not.toBeInTheDocument();
-      });
-      expect(screen.queryByTestId("label-enrollment-days")).toHaveTextContent(
-        "1 day",
-      );
     });
 
     it("handles submission with server API error", async () => {
